refactored according to CR
