+++ BruteShark/BruteSharkCli/Cli/CliShell.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using PcapAnalyzer;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -10,18 +11,53 @@ namespace BruteSharkCli
     class CliShell
     {
         public string Seperator { get; set; }
+        private ulong _tcpPacketsCount;
+        private ulong _udpPacketsCount;
+        private int _tcpSessionsCount;
+        private int _udpStreamsCount;
+        private object _printingLock;
         private List<CliShellCommand> _commands;
         private bool _exit;
+        private HashSet<PcapAnalyzer.NetworkPassword> _passwords;
+        private HashSet<PcapAnalyzer.NetworkHash> _hashes;
+        private List<string> _files;
+        private HashSet<PcapAnalyzer.NetworkConnection> _connections;
+        private PcapAnalyzer.Analyzer _analyzer;
+        private PcapProcessor.Processor _processor;
 
-        public CliShell(string seperator = ">")
+        public CliShell(PcapAnalyzer.Analyzer analyzer, PcapProcessor.Processor processor, string seperator = ">")
         {
+            _printingLock = new object();
+            _tcpPacketsCount = 0;
+            _udpPacketsCount = 0;
+            _udpStreamsCount = 0;
+            _tcpSessionsCount = 0;
+            PcapProcessor.Processor _processor = processor;
+            PcapAnalyzer.Analyzer _analyzer = analyzer;
+            _analyzer.ParsedItemDetected += OnParsedItemDetected;
+            _processor.TcpPacketArived += (s, e) => this.UpdateTcpPacketsCount();
+            _processor.UdpPacketArived += (s, e) => this.UpdateUdpPacketsCount();
+            _processor.TcpSessionArrived += (s, e) => this.UpdateTcpSessionsCount();
+            _processor.UdpSessionArrived += (s, e) => this.UpdateUdpStreamsCount();
+            HashSet<PcapAnalyzer.NetworkConnection> _connections = new HashSet<PcapAnalyzer.NetworkConnection>();
+            List<string> _files = new List<string>();
+            HashSet<PcapAnalyzer.NetworkPassword> _passwords = new HashSet<PcapAnalyzer.NetworkPassword>();
+            HashSet<PcapAnalyzer.NetworkHash> _hashes = new HashSet<PcapAnalyzer.NetworkHash>();
             this.Seperator = seperator;
             this._commands = new List<CliShellCommand>();
+            AddCommand(new CliShellCommand("add-file", p => AddFile(p), "Add file to analyze. Usage: add-file <FILE-PATH>"));
+            AddCommand(new CliShellCommand("start", p => StartAnalyzing(), "Start analyzing"));
+            AddCommand(new CliShellCommand("show-passwords", p => PrintPasswords(), "Print passwords."));
+            AddCommand(new CliShellCommand("show-modules", p => PrintModules(), "Print modules."));
+            AddCommand(new CliShellCommand("show-hashes", p => PrintHashes(), "Print Hashes"));
+            AddCommand(new CliShellCommand("show-networkmap", p => PrintNetworkMap(), "Prints the network map as a json string. Usage: show-networkmap"));
+            AddCommand(new CliShellCommand("export-hashes", p => Utilities.ExportHashes(p, _hashes), "Export all Hashes to Hascat format input files. Usage: export-hashes <OUTPUT-DIRECTORY>"));
+            AddCommand(new CliShellCommand("export-networkmap", p => Utilities.ExportNetworkMap(p, _connections), "Export network map to a json file for neo4j. Usage: export-networkmap <OUTPUT-file>"));
 
             // Add the help command
             this.AddCommand(new CliShellCommand(
                 "help",
-                 param => this.PrintCommandsWithDescription(), 
+                 param => this.PrintCommandsWithDescription(),
                  "Print help menu"));
 
             // Add the exit command
@@ -29,6 +65,14 @@ namespace BruteSharkCli
                 "exit",
                  param => this._exit = true,
                  "Exit CLI"));
+
+        }
+        private void PrintModules()
+        {
+            foreach (string module in this._analyzer.AvailableModulesNames)
+            {
+                Console.WriteLine($" - {module}");
+            }
         }
 
         public void AddCommand(CliShellCommand cliShellCommand)
@@ -43,6 +87,9 @@ namespace BruteSharkCli
 
         internal void Start()
         {
+
+            Utilities.PrintBruteSharkAsciiArt();
+            RunCommand("help");
             _exit = false;
 
             do
@@ -51,8 +98,6 @@ namespace BruteSharkCli
             }
             while (!_exit);
         }
-
-     
 
         private void HandleUserInput()
         {
@@ -89,6 +134,99 @@ namespace BruteSharkCli
             return legalInput;
         }
 
+        private void PrintPasswords()
+        {
+            _passwords.ToDataTable().Print();
+        }
 
-    }
+        private void PrintHashes()
+        {
+            this._hashes.ToDataTable(itemLengthLimit: 15).Print();
+        }
+
+        private void PrintNetworkMap()
+        {
+            Console.WriteLine(NetwrokMapJsonExporter.GetNetworkMapAsJsonString(this._connections.ToList()));
+        }
+
+        private void StartAnalyzing()
+        {
+            _processor.ProcessPcaps(this._files);
+            Console.SetCursorPosition(0, Console.CursorTop + 5);
+        }
+
+        private void UpdateTcpSessionsCount()
+        {
+            ++_tcpSessionsCount;
+            UpdateAnalyzingStatus();
+        }
+        private void UpdateUdpStreamsCount()
+        {
+            ++_udpStreamsCount;
+            UpdateAnalyzingStatus();
+        }
+
+            private void UpdateTcpPacketsCount()
+            {
+                if (++_tcpPacketsCount % 10 == 0)
+                {
+                    UpdateAnalyzingStatus();
+                }
+            }
+            private void UpdateUdpPacketsCount()
+            {
+                if (++_udpPacketsCount % 10 == 0)
+                {
+                    UpdateAnalyzingStatus();
+                }
+            }
+            private void UpdateAnalyzingStatus()
+            {
+                lock (_printingLock)
+                {
+                    Console.ForegroundColor = ConsoleColor.Blue;
+                    Console.WriteLine($"\r[+] Packets Analyzed: {_tcpPacketsCount + _udpPacketsCount}, " + $"TCP: {_tcpPacketsCount} " + $"UDP: {_udpPacketsCount}");
+                    Console.WriteLine($"\r[+] TCP Sessions Analyzed: {_tcpSessionsCount}" + $" UDP Streams Analyzed: {_udpStreamsCount}");
+                    Console.WriteLine($"\r[+] Passwords Found: {_passwords.Count}");
+                    Console.WriteLine($"\r[+] Hashes Found: {_hashes.Count}");
+                    Console.WriteLine($"\r[+] Network Connections Found: {_connections.Count}");
+                    Console.SetCursorPosition(0, Console.CursorTop - 5);
+                    Console.ForegroundColor = ConsoleColor.White;
+                }
+            }
+        private void AddFile(string filePath)
+        {
+            if (File.Exists(filePath))
+            {
+                _files.Add(filePath);
+            }
+            else
+            {
+                Console.WriteLine("File does not exist.");
+            }
+
+        }
+        private void OnParsedItemDetected(object sender, ParsedItemDetectedEventArgs e)
+        {
+            if (e.ParsedItem is PcapAnalyzer.NetworkPassword)
+            {
+                _passwords.Add(e.ParsedItem as PcapAnalyzer.NetworkPassword);
+            }
+            if (e.ParsedItem is PcapAnalyzer.NetworkHash)
+            {
+                _hashes.Add(e.ParsedItem as PcapAnalyzer.NetworkHash);
+            }
+            if (e.ParsedItem is PcapAnalyzer.NetworkConnection)
+            {
+                var networkConnection = e.ParsedItem as NetworkConnection;
+                _connections.Add(networkConnection);
+            }
+
+            UpdateAnalyzingStatus();
+        }
+       
+    } 
 }
+
+
+
