+++ BruteShark/PcapAnalyzer/Modules/FilesExtractingModule/FileExtactingModule.cs
@@ -10,8 +10,7 @@ namespace PcapAnalyzer
 {
     public class FileExtactingModule : IModule
     {
-        public string Name => "File Extracintg";
-        public string CliName => "FileExtracting";
+        public string Name => "File Extracting";
         public event EventHandler<ParsedItemDetectedEventArgs> ParsedItemDetected;
 
         // File signitures examples:
