+++ BruteShark/PcapAnalyzer/Modules/NetworkMapModule/NetworkMapModule.cs
@@ -7,7 +7,6 @@ namespace PcapAnalyzer
     class NetworkMapModule : IModule
     {
         public string Name => "Network Map";
-        public string CliName => "NetworkMap";
 
         public event EventHandler<ParsedItemDetectedEventArgs> ParsedItemDetected;
